+++ .github/workflows/main.yml
@@ -39,6 +39,7 @@ jobs:
           --max-complexity=10 \
           --max-line-length=110 \
           --statistics \
+          --ignore WPS115,WPS601,WPS211,WPS230,WPS404
           # --ignore WPS404,WPS407,WPS115,WPS430,WPS458,E402,WPS229,E501,WPS226,WPS501,WPS234,WPS601,WPS211,WPS230,WPS336,WPS102,WPS420,WPS110,WPS463,WPS111,WPS300,W503
           # --exit-zero
